refactor(css): establish consistent component folder structure with co-located CSS
Each component now lives in its own named folder following the convention:
- component-name/component-name.svelte
- component-name/component-name.css

Components refactored:
- OKRIndex → okr-index/
- ObjectiveCard → objective-card/
- KeyResultRow → key-result-row/
- KeyResultModal → key-result-modal/
- ObjectiveModal → objective-modal/
- ProgressRing → progress-ring/

Component-specific CSS extracted from app.css into co-located .css files,
imported via @import in each component's <style> block. app.css now contains
only global styles: design tokens, resets, button variants, and form elements.

Fixes kogeletey/kefine#38
