+++ src/lib/components/okrs/key-result-modal/key-result-modal.svelte
@@ -211,3 +211,7 @@
     </form>
   </dialog>
 </section>
+
+<style>
+  @import './key-result-modal.css';
+</style>
