+++ src/lib/components/okrs/key-result-row/key-result-row.css
@@ -1,0 +1,120 @@
+/* li.kr-row: key result list item */
+.kr-row {
+  display: flex;
+  flex-direction: column;
+  gap: var(--okr-space-2);
+  list-style: none;
+  padding: var(--okr-space-3) 0;
+  border-bottom: 1px solid var(--okr-color-border-light);
+}
+
+.kr-row:last-child {
+  border-bottom: none;
+}
+
+.kr-header {
+  display: flex;
+  align-items: center;
+  gap: var(--okr-space-2);
+}
+
+/* kr-type-badge: target type label (custom element) */
+kr-type-badge {
+  display: inline-block;
+  font-size: var(--okr-font-size-xs);
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+  flex-shrink: 0;
+}
+
+/* kr-title: key result title (custom element) */
+kr-title {
+  display: block;
+  font-size: var(--okr-font-size-sm);
+  font-weight: 600;
+  color: var(--okr-color-text);
+  flex: 1;
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.kr-description {
+  display: block;
+  font-size: var(--okr-font-size-xs);
+  color: var(--okr-color-muted);
+}
+
+.kr-progress {
+  display: flex;
+  align-items: center;
+  gap: var(--okr-space-2);
+}
+
+/* kr-progress-bar: progress track (custom element) */
+kr-progress-bar {
+  display: block;
+  flex: 1;
+  height: 8px;
+  background: #e5e7eb;
+  border-radius: var(--okr-radius-full);
+  overflow: hidden;
+}
+
+/* kr-progress-fill: progress fill (custom element) */
+kr-progress-fill {
+  display: block;
+  height: 100%;
+  border-radius: var(--okr-radius-full);
+  transition: width 400ms ease;
+}
+
+.kr-progress-pct {
+  display: inline-block;
+  font-size: var(--okr-font-size-xs);
+  font-weight: 600;
+  color: var(--okr-color-muted);
+  min-width: 2.5rem;
+  text-align: right;
+}
+
+.kr-values {
+  display: flex;
+  font-size: var(--okr-font-size-sm);
+  color: var(--okr-color-muted);
+  gap: var(--okr-space-1);
+}
+
+/* kr-current: current value highlight (custom element) */
+kr-current {
+  display: inline;
+  font-weight: 700;
+  color: var(--okr-color-text);
+}
+
+/* kr-sep: separator (custom element) */
+kr-sep {
+  display: inline;
+  color: var(--okr-color-border);
+}
+
+/* kr-target: target value (custom element) */
+kr-target {
+  display: inline;
+}
+
+.kr-edit-form {
+  display: flex;
+  flex-direction: column;
+  gap: var(--okr-space-2);
+  padding: var(--okr-space-2);
+  background: var(--okr-color-bg);
+  border-radius: var(--okr-radius-md);
+}
+
+.kr-edit-actions {
+  display: flex;
+  gap: var(--okr-space-2);
+}
