+++ src/lib/components/okrs/key-result-row/key-result-row.svelte
@@ -119,3 +119,7 @@
     </fieldset>
   {/if}
 </li>
+
+<style>
+  @import './key-result-row.css';
+</style>
