+++ src/lib/components/okrs/objective-card/objective-card.css
@@ -1,0 +1,115 @@
+/* article.okr-card: objective card */
+.okr-card {
+  display: flex;
+  flex-direction: column;
+  background: var(--okr-color-bg-card);
+  border-radius: var(--okr-radius-xl);
+  box-shadow: var(--okr-shadow-md);
+  overflow: hidden;
+  transition: box-shadow 250ms ease, transform 250ms ease;
+}
+
+.okr-card:hover {
+  box-shadow: var(--okr-shadow-lg);
+  transform: scale(1.01);
+}
+
+/* TAC attribute variation: status */
+.okr-card[data-status="archived"] {
+  opacity: 0.65;
+}
+
+.okr-card-header {
+  display: block;
+  padding: var(--okr-space-4) var(--okr-space-6) var(--okr-space-3);
+  border-bottom: 1px solid var(--okr-color-border-light);
+}
+
+.okr-card-title-row {
+  display: flex;
+  align-items: center;
+  gap: var(--okr-space-3);
+  margin-bottom: var(--okr-space-2);
+}
+
+/* okr-card-meta: title + quarter (custom element) */
+okr-card-meta {
+  display: block;
+  flex: 1;
+  min-width: 0;
+}
+
+okr-card-meta h3 {
+  font-size: var(--okr-font-size-base);
+  font-weight: 700;
+  color: var(--okr-color-text);
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  margin: 0;
+}
+
+/* okr-quarter: quarter/year label (custom element) */
+okr-quarter {
+  font-size: var(--okr-font-size-xs);
+  color: var(--okr-color-muted);
+}
+
+/* okr-status-badge: status pill (custom element) */
+okr-status-badge {
+  display: inline-block;
+  font-size: var(--okr-font-size-xs);
+  font-weight: 600;
+  text-transform: capitalize;
+  padding: 0.15rem 0.5rem;
+  border-radius: var(--okr-radius-full);
+  flex-shrink: 0;
+}
+
+.okr-card-description {
+  display: -webkit-box;
+  font-size: var(--okr-font-size-sm);
+  color: var(--okr-color-muted);
+  -webkit-line-clamp: 3;
+  line-clamp: 3;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  margin: 0;
+}
+
+.okr-card-body {
+  display: block;
+  padding: var(--okr-space-3) var(--okr-space-6);
+  flex: 1;
+}
+
+.okr-kr-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: var(--okr-space-2);
+}
+
+/* okr-kr-count: key results count label (custom element) */
+okr-kr-count {
+  display: inline-block;
+  font-size: var(--okr-font-size-xs);
+  font-weight: 600;
+  color: var(--okr-color-muted);
+}
+
+.okr-kr-empty {
+  display: block;
+  font-size: var(--okr-font-size-sm);
+  color: var(--okr-color-muted);
+  text-align: center;
+  padding: var(--okr-space-3) 0;
+}
+
+.okr-card-footer {
+  display: flex;
+  gap: var(--okr-space-2);
+  padding: var(--okr-space-2) var(--okr-space-6);
+  border-top: 1px solid var(--okr-color-border-light);
+  background: var(--okr-color-bg);
+}
