+++ src/lib/components/okrs/progress-ring/progress-ring.css
@@ -1,0 +1,4 @@
+/* .progress-arc: animated SVG progress arc */
+.progress-arc {
+  transition: stroke-dashoffset 500ms ease-out, stroke 300ms ease;
+}
